+++ CHANGELOG.md
@@ -155,6 +155,8 @@
   add extra metadata passed in from hints in knife-linode
 * Update mime-types dependency
 
+* Upadate Rackspace plugin to work on Windows platform.
+
 ## Release 8.0.0
 
 * [**sawanoboly**](https://github.com/sawanoboly)